+++ src/components/Main.vue
@@ -63,6 +63,7 @@ import CoordinatesInput from "../components/CoordinatesInput";
 import HUCInput from "../components/HUCInput";
 import ParamSelect from "../components/ParamSelect";
 import { mapState } from "vuex";
+import { notify } from "../notifications.js";
 
 /*global  tableau:true*/
 
@@ -97,12 +98,23 @@ export default {
   },
   methods: {
     /*
+      Warns users on standalone browsers that they need Tableau to proceed with data collection.
+    */
+    browserWarning: function() {
+      if (tableau.platformVersion == undefined) {
+        notify(
+          "The NWIS Tableau Web Data Connector must be accessed from Tableau desktop or Tableau server!"
+        );
+      }
+    },
+    /*
             This function is triggered when the user presses the button to confirm their query. 
             This closes the Web Data Connector interface.
         */
     requestData: function() {
+      this.browserWarning();
       if (!this.loadedStateData) {
-        alert(
+        notify(
           "The page is still loading: please retry this action in a moment!"
         );
         return;
@@ -161,6 +173,7 @@ export default {
   mounted: function() {
     this.$nextTick(function() {
       this.fetchData();
+      this.browserWarning();
     });
   },
   watch: {