+++ CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added input for selection of multiple site types by user
 - Added support for semantic versioning into project
 - Jenkinsfile.build will now allow users to automatically increment version numbers in package.json and add git tags to builds
-
+- Added Tableau Desktop Compatible alerts
 
 ### Changed